libprefetch: Remove `start` command
This command evaluates the current state of device
and sets either `record` or `replay` property depending
on the result.

As these checks are now performed by its respective
commands by their own, `start` is now obsolete and can
be removed.

Bug: 380766679
Test: Build
Test: on cuttlefish, observe that record only starts when
the same condition originally checked in start is met.
Test: on cuttlefish, observe that replay only starts when
the same condition originally checked in start is met.

Change-Id: Idb791187b1c8d3c69a1d7f662a4b74413f6819a2