libtrusty: Remove status byte workaround in TIPC vsock connect
This change removes the logic that reads a single status byte
from the vsock-TIPC forwarder after sending the service name, as
the sync byte sending is removed in the guest's vsock bridge.

This status byte was originally a workaround to ensure the
vsock-TIPC forwarder had processed the service name and was ready
for data before the client sent more.

Bug: 406418102
Test: atest vts_treble_vintf_trusted_hal_test
Change-Id: I41dbb13402ccade5a0853e4d40517c164da8d7ec