init: Ignore `BIONIC_SIGNAL_PROFILER` before `execv`
There is a race condition where Perfetto may send its signal to start
tracing for a process before it has installed appropriate signal
handlers.  In that situation, the process gets killed because the
default signal disposition for this particular signal is to kill the
process.

When a process is launched via fork-exec, the window where this can
happen is between the exec, which removes signal handlers and resets the
signal dispositions for those signals, and the libc init, which installs
new signal handlers.  This behavior of exec is specified by POSIX, and
so makes it difficult to solve for this race condition generally within
bionic's implementation of exec, as there are many perfectly legal ways
to leak attempts at signal blocking into user-visible code and break the
expectations of POSIX.

This change proposes a narrower solution.  Rather than solve for the
general case, we can solve for a specific case that we have seen occur
many times in the field: when bug reports are collected, the dumpstate
service is launched by init, and if Perfetto tracing is running
systemwide, then dumpstate may be killed and the bug report collection
will fail.  (This appears to happen more frequently on watches, where
the CPUs are generally slower than other form factors, and so the window
for the race condition is a bit wider).  As a solution for this case, I
set the signal to ignore prior to the exec when init launches a service.
POSIX specifies that the signal will continue to be ignored across the
exec.  On the other side, we expect that bionic would then install the
signal handler as usual, thus enabling it again.

FIXED_READ_ONLY_FLAG_REASON=NATIVE_FLAG_PRE_METADATA_MOUNT this flag needs to be available early
Bug: 357736486, 151835887
Flag: com.android.init.flags.ignore_bionic_signal_profiler_before_exec
Test: strace to verify, see b/357736486
Change-Id: Idbfd62716adccfcaa20d7a03ff15db6be15bdc6e