Partially reapply "ashmem: Change requirements for using memfd"
This reverts commit 4532e8a2ff39e4b81912ee0f0777ba03c7b0706b.

Commit 4532e8a2ff39 ("ashmem: Change requirements for using memfd") was
reverted because it caused several codec processes to crash on
upgrading devices. This was because some of the memfd related syscalls
(i.e. fcntl() and fsetxattr()) weren't allow-listed in the seccomp
filters for the codec processes.

While the fix would just be to simply allow-list the syscalls, the
codec processes and their seccomp filters were part of the vendor
image, meaning that using memfd would require vendor software to
update as a result of an Android platform upgrade, which violates
the vendor requirements freeze guidelines.

Therefore, use memfd only if the vendor image's API level is at least
202604, which is the API level for Android 17. The rest of the patch
is the same as the first version that was initially merged.

Bug: 111903542
Change-Id: I7c9c434680c3684956facf87033b427736310cc0
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>