Fix prefetch on first boot when fingerprint is missing
The http://ag/36506950 makes prefetch record not run in the OTA boot,
and build fingerprint file is not written /metadata during OTA boot. If
the metadata is wiped, the build fingerprint will never be created.

Instead of returning false when the fingerprint file is missing, we just
use the default empty string value as fingerprint reference.

Bug: 452520774
Bug: 455809368
Test: CQ and abtd A/B test run
Change-Id: I4f06cc83e3146ff9af65fe5ed4d6edf48cd00ca9