ashmem: Change requirements for using memfd
Instead of relying on the use_memfd system property to determine when
to use memfd, use memfd under the following conditions:

1. The kernel supports the memfd_class SELinux policy capability
2. The vendor API level of the device is 202604 or higher
3. The application targets SDK version 37 or higher

The check for kernel support for the memfd_class SELinux policy
is needed to avoid using memfd on older kernels that do not have
support for memfd_class. This is only needed for Cuttlefish x86
for WearOS, which is limited to kernel 6.1 as it cannot upgrade
to a newer kernel due to limitations in the x86 networking
drivers.

The vendor API level requirement ensures that the memfd_class
capability is enabled, as it is required for devices launching
with vendor API level 202604 or higher.

The application target SDK version requirement allows backwards
compatibility with older applications.

Bug: 111903542
Change-Id: Iccfc0e6e126bdcdf15ec76fa54ff6313cd003ed9
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>