init: cleanup perf_event_open LSM support detection
This patch removes the runtime check that was used to detect kernel
support for selinux hooks in the perf_event_open syscall, and hardcodes
the system-wide perf_event_paranoid sysctl to -1.

The hooks were added upstream in v5.5 (2019). Earlier kernels are now
considered EOL in Android [1], so we now assume kernel support without
any further testing.

[1] https://source.android.com/docs/core/architecture/kernel/android-common#support-lifetimes

Bug: 383134654
Bug: 483294088
Change-Id: I78a8daa832a996968015bd2bc8122852ea89f82a