Add RunWithNoCrash helper to FastbootDriver
FastbootDriver's methods would crash on error today, making
the return value useless. As these methods eithe return success
or do not return at all. Add a RunWithNoCrash() helper, which
takes a functor and runs it with crash on error disabled. This
allows the caller to gracefully handle commands which are allowed
to fail.

Test: th
Bug: 459913876
Change-Id: I17ae4a269f9273613a0dbbd017b903030db2611a