storage: better handling of driver soft-retries
The retry logic for SECURITY IN and OUT commands assumes that a retry is
caused by something like the device POWER ON or RESET condition, and
hence the retry can only be for the first command.

A previous case added retries for a driver case, DID_REQUEUE.  In this
case retrying the command should be allowed as it is not a device
condition.

Bug: 455221920
Test: boot to home
Test: reboot stress test
Change-Id: I83a747e6c7e11ce14ff08f7eb3fd476ad909452c