init: Force an unlink of snapuserd in the vendor ramdisk if using init_boot.
Android 12's first-stage ramdisk had no mechanism to split files between
the system and vendor sides of the Treble boundary. Android 13 fixed
this with the init_boot partition.

For Android 12 launch devices, snapuserd was considered to be vendor
code in the ramdisk, and thus participated in vendor freeze. This vendor
version is no longer supported, but the code for it in init leads to
issues if vendor_boot accidentally contains an additional copy of
snapuserd.

To work around this, unlink the vendor copy of snapuserd if it exists in
addition to the init_boot variant.

Bug: 465647027
Test: th
Change-Id: Ia226cc6b0efc47f3cc1f14c2f0c7129f2dea35f1