toolbox: Move modprobe out of the toolbox
Modprobe is currently built as a part of the toolbox toolkit which
execution workflow is a bit tangled and the build rules are difficult to
extend to include the new variations - bootstraped binaries. The
statically linked modprobe binary is required to load the kernel
modules in the pre-APEX environment by the usermode helper tool.

Move the modprobe binary out of the toolbox and place it in a separate
directory with it's own Android.bp file declaring the modprobe as
bootstrapped by default. Provide required variations - vendor, ramdisk
and recovery as well and add it to the shell_and_utilities accordingly.

Bug: 429298818
Test: Verify that requested modules are loaded with the modprobe
bootstrap binary.
Test: Presubmit.
Test: Verify the modprobe binaries are built.

Change-Id: Ie2569042221507ec163a7e3a6afe08bc0061e58e